backend/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mixer==7.2.2
pillow==10.3.0
ruff==0.3.7